requirements.txt
pip~=24.2
Django~=5.1
asgiref~=3.8.1
sqlparse~=0.5.1
setuptools~=65.5.1
djangorestframework~=3.15.2